fix(proto): give the preview lifetimes the width their wire field has
PREVIEW_SESSION_LIFETIME_SECS was u32 while the IPC field carrying it is
u16, so the VM owner's binary did not compile. The constants now carry
the type they are sent as.

+++ crates/capsem-proto/src/lib.rs
@@ -103,9 +103,9 @@ pub enum PublicationAccess {
 /// it and the confined router that strips it before the workload sees it.
 pub const PREVIEW_COOKIE: &str = "capsem_preview";
 /// How long a single-use browser bootstrap token may be exchanged.
-pub const PREVIEW_BOOTSTRAP_LIFETIME_SECS: u32 = 30;
+pub const PREVIEW_BOOTSTRAP_LIFETIME_SECS: u16 = 30;
 /// How long an exchanged preview session admits connections.
-pub const PREVIEW_SESSION_LIFETIME_SECS: u32 = 15 * 60;
+pub const PREVIEW_SESSION_LIFETIME_SECS: u16 = 15 * 60;
 
 #[derive(Serialize, Deserialize, Debug, Clone, Copy, PartialEq, Eq)]
 #[serde(rename_all = "snake_case")]
